Fix card for new layout
