+++ client/my-sites/customer-home/cards/education/educational-content/index.jsx
@@ -2,11 +2,13 @@
  * External dependencies
  */
 import React from 'react';
+import classnames from 'classnames';
 import { isDesktop } from '@automattic/viewport';
 
 /**
  * Internal dependencies
  */
+import ExternalLink from 'components/external-link';
 import InlineSupportLink from 'components/inline-support-link';
 import Gridicon from 'components/gridicon';
 
@@ -24,18 +26,30 @@ const EducationalContent = ( { title, description, links, illustration } ) => {
 					{ description }
 				</p>
 				<div className="educational-content__links">
-					{ links.map( ( { postId, url, text, icon, tracksEvent, statsName } ) => (
-						<div className="educational-content__link" key={ postId }>
+					{ links.map( ( { postId, url, externalLink, text, icon, tracksEvent, statsName } ) => (
+						<div
+							className={ classnames( 'educational-content__link', {
+								'is-external-link': externalLink,
+							} ) }
+							key={ url }
+						>
 							{ icon && <Gridicon icon={ icon } size={ 18 } /> }
-							<InlineSupportLink
-								supportPostId={ postId }
-								supportLink={ url }
-								showIcon={ false }
-								text={ text }
-								tracksEvent={ tracksEvent }
-								statsGroup="calypso_customer_home"
-								statsName={ statsName }
-							/>
+							{ postId && (
+								<InlineSupportLink
+									supportPostId={ postId }
+									supportLink={ url }
+									showIcon={ false }
+									text={ text }
+									tracksEvent={ tracksEvent }
+									statsGroup="calypso_customer_home"
+									statsName={ statsName }
+								/>
+							) }
+							{ externalLink && (
+								<ExternalLink href={ url } icon>
+									{ text }
+								</ExternalLink>
+							) }
 						</div>
 					) ) }
 				</div>
