+++ client/my-sites/customer-home/cards/education/educational-content/style.scss
@@ -23,7 +23,7 @@
 			margin-bottom: 0;
 		}
 
-		.gridicon {
+		&:not( .is-external-link ) .gridicon {
 			vertical-align: text-bottom;
 			margin-right: 0.5rem;
 			fill: var( --color-neutral );
